Moved key into tokens instead of redeem
