+++ gradle.properties
@@ -2,7 +2,7 @@
 ## # This is actually the "correct" place to define the version for the project.
 ## # Used within build.gradle with ${project.version}.
 ## # Can be overridden on the command line: gradle -Pversion=3.9.13-alpha.3
-version=4.0.1
+version=4.1.0
 
 ## org.gradle.warning.mode=(all,none,summary)
 org.gradle.warning.mode=all
